+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - post-download cutout creation, store in hdf5 files
 - retry logic for downloads
 - plotting routine for cutouts (unionsdata plot)
+- integrate config validation into unionsdata edit and remove unionsdata validate
 
 ## [0.1.1] - 2025-10-31